# ✨ v3 (enhancement): add `[]byte` support to `utils.EqualFold`