@efectn 2022-08-19
I think it's good idea to use generics for Fiber utils.