@Trim21 2022-08-19
I don't why, But I find that using `[]byte` (`BytesEqualFold(b, s []byte) bool`) is even slower than generic `byteSeq`...

```text
goos: windows
goarch: amd64
pkg: github.com/gofiber/fiber/v3/utils
cpu: AMD Ryzen 7 5800X 8-Core Processor
Benchmark_BytesEqualFold
Benchmark_BytesEqualFold/fiber
Benchmark_BytesEqualFold/fiber-16               29250812                40.94 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-16               29241974                40.81 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-16               29241760                40.97 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-16               28878572                40.91 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-bytes-equal-fold
Benchmark_BytesEqualFold/fiber-bytes-equal-fold-16              24204332                49.42 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-bytes-equal-fold-16              23744884                49.71 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-bytes-equal-fold-16              24976584                49.38 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/fiber-bytes-equal-fold-16              23978371                49.14 ns/op            0 B/op          0 allocs/op
Benchmark_BytesEqualFold/default
Benchmark_BytesEqualFold/default-16                              9628075               123.2 ns/op             0 B/op          0 allocs/op
Benchmark_BytesEqualFold/default-16                              9762962               123.7 ns/op             0 B/op          0 allocs/op
Benchmark_BytesEqualFold/default-16                              9604748               123.5 ns/op             0 B/op          0 allocs/op
Benchmark_BytesEqualFold/default-16                              9665220               123.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v3/utils       15.386s
```