@ReneWerner87 2022-08-19
by the way i like such changes when the functions can do more and the performance stays the same

now can't we remove the old EqualFold function for bytes ?

we can also try to replace other old functions with many parameters with configuration items instead of 3-4 arguments so that they are extensible