@ReneWerner87 2022-08-19
https://github.com/gofiber/fiber/blob/v3-beta/utils/bytes.go#L59