@Trim21 2022-08-19
Oh, I didn't find there is a `EqualFoldBytes` for bytes... I'll remove it in https://github.com/gofiber/fiber/pull/2030